Initial work - Projects - don't cancel the marketplace/search request when loading the page. #5780
