+++ ui/app/src/components/CreateSiteDialog/translations.ts
@@ -1,0 +1,106 @@
+/*
+ * Copyright (C) 2007-2022 Crafter Software Corporation. All Rights Reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as published by
+ * the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { defineMessages } from 'react-intl';
+
+export const messages = defineMessages({
+  privateBlueprints: {
+    id: 'createSiteDialog.privateBlueprints',
+    defaultMessage: 'Private Blueprints'
+  },
+  marketplace: {
+    id: 'common.marketplace',
+    defaultMessage: 'Marketplace'
+  },
+  publicMarketplaceBlueprints: {
+    id: 'createSiteDialog.publicMarketplaceBlueprints',
+    defaultMessage: 'Public Marketplace Blueprints'
+  },
+  back: {
+    id: 'common.back',
+    defaultMessage: 'Back'
+  },
+  noMarketplaceBlueprints: {
+    id: 'createSiteDialog.noMarketplaceBlueprints',
+    defaultMessage: 'No Marketplace Blueprints Were Found'
+  },
+  changeQuery: {
+    id: 'createSiteDialog.changeQuery',
+    defaultMessage: 'Try changing your query or browse the full catalog.'
+  },
+  creatingSite: {
+    id: 'createSiteDialog.creatingSite',
+    defaultMessage: 'Creating Project'
+  },
+  pleaseWait: {
+    id: 'createSiteDialog.pleaseWait',
+    defaultMessage: 'Please wait while your project is being created.'
+  },
+  createInBackground: {
+    id: 'createSiteDialog.createInBackground',
+    defaultMessage: 'Create in Background'
+  },
+  dialogCloseTitle: {
+    id: 'createSiteDialog.dialogCloseTitle',
+    defaultMessage: 'Confirm Close'
+  },
+  dialogCloseMessage: {
+    id: 'createSiteDialog.dialogCloseMessage',
+    defaultMessage: 'Data entered in the form would be lost upon closing.'
+  },
+  gitBlueprintName: {
+    id: 'createSiteDialog.gitBlueprintName',
+    defaultMessage: 'Remote Git Repository'
+  },
+  gitBlueprintDescription: {
+    id: 'createSiteDialog.gitBlueprintDescription',
+    defaultMessage: 'Create a new project based on a CrafterCMS project in an existing, remote git repository.'
+  },
+  createSite: {
+    id: 'createSiteDialog.createSite',
+    defaultMessage: 'Create Project'
+  },
+  review: {
+    id: 'createSiteDialog.review',
+    defaultMessage: 'Review'
+  },
+  finish: {
+    id: 'createSiteDialog.finish',
+    defaultMessage: 'Finish'
+  },
+  nameAndDescription: {
+    id: 'createSiteDialog.nameAndDescription',
+    defaultMessage: 'Name and describe your project'
+  },
+  reviewSite: {
+    id: 'createSiteDialog.reviewSite',
+    defaultMessage: 'Review set up summary and create your project'
+  },
+  showIncompatible: {
+    id: 'createSiteDialog.showIncompatible',
+    defaultMessage: 'Show Incompatible Plugins'
+  },
+  marketplaceUnavailable: {
+    id: 'createSiteDialog.marketplaceUnavailable',
+    defaultMessage: 'CrafterCMS Marketplace is unavailable at this time'
+  },
+  retry: {
+    id: 'words.retry',
+    defaultMessage: 'Retry'
+  }
+});
+
+export default messages;
